fix6

+++ src/Controller/FatchipComputopPaypalExpress.php
@@ -335,9 +335,11 @@ class FatchipComputopPayPalExpress extends FrontendController
         //this condition if true indicate the user has been tmp created during the PaypalExpress createOrder action
         if ($this->stringStartWith($oUser->oxuser__oxusername->value, 'PAYPAL_TMP_USER')) {
             $sResponseEmailId = $oUser->getIdByUserName($oResponse->getEMail());
+            if (!empty($sResponseEmailId)) {
                 $oOrder->oxorder__oxuserid = new Field($sResponseEmailId);
                 $oUser->delete();
                 $oUser->load($sResponseEmailId);
+            }
                 $notApplicableFields = [
                     'oxuser__oxustid'  => '',
                     'oxuser__oxcompany' => '',
